vhost_user_backend: implement support for asynchronous completions
This allows crate users to receive completions on a separate
thread. For each vring, a channel and a completion worker are
created. One endpoint of the channel is polled by the worker, while
the other is send to the backend through "handle_event".

The backend can then use that endpoint to signal when a completion
for a request must be run. This relieves the backend from the need to
complete the request in the "handle_event" path and, in case those
requests are being processed in parallel, allows to serialize the
completions to reduce the pressure on the vring lock.

As an example of the above, serializing the completions for parallel
requests in vhost_user_fs raised the performance with a fio
test (ioengine=libaio, iodepth=64, numjobs=1) over a cached file from
~40K to ~70K. Other tests with smaller iodepths showed a lower, but
still significant, improvement.

Signed-off-by: Sergio Lopez <slp@redhat.com>