vhost_user_fs: Process requests in parallel with a thread pool
This change enables vhost_user_fs to process multiple requests in
parallel by scheduling them into a ThreadPool (from the Futures
crate). Completions are processed sequentially by using the
"handle_completion" and "completion workers" instruments introduced in
vhost_user_backend by a previous commit.

Parallelism on a single file is limited by the nature of the operation
executed on it. A recent commit replaced the Mutex that protects the
File within HandleData with a RwLock, to allow some operations (at
this moment, only "read" and "write") to proceed in parallel by
acquiring a read lock.

This strategy gives a performance boost to virtually all
non-synchronous workloads, without having a measurable impact on
the synchronous ones.

Signed-off-by: Sergio Lopez <slp@redhat.com>